refactor(services): PR-3.4 models/ + preprocess/ 子包搬迁 — 4 文件
models/:
  services/model_downloader.py → services/models/downloader.py（**整文件搬迁，4-way 拆留独立 commit 在 PR-3 末尾**）

preprocess/:
  studio/preprocess.py            → studio/services/preprocess/core.py（**顶层业务模块搬入 services/**）
  services/preprocess_manifest.py → services/preprocess/manifest.py
  services/duplicate_finder.py    → services/preprocess/duplicates.py

回归：211 passed（含 preprocess / model_downloader / preprocess_manifest 测试）
